responsive design
responsive design,
added toaster to delete button in all tables

+++ src/app/super-admin-module/cateogry/cateogry.component.ts
@@ -3,6 +3,7 @@ import { BlogsService } from '../../api/api-services/blogs.service';
 import { Router } from '@angular/router';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { HttpErrorResponse } from '@angular/common/http';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-cateogry',
@@ -61,24 +62,58 @@ export class CateogryComponent {
   }
 
   // Delete Category
+  // deleteCategory(id: any) {
+  //   this.blogsService.deleteCategoryList(id).subscribe({
+  //     next: (res: any) => {
+  //       this._snackBar.open(res.message, 'Close', {
+  //         duration: 3000,
+  //         verticalPosition: 'bottom',
+  //         horizontalPosition: 'center'
+  //       });
+  //       if (res.status === 200) {
+  //         this.getCategoryList()
+  //       }
+  //     }, error: (err: HttpErrorResponse) => {
+  //       this._snackBar.open(err.statusText, 'Close', {
+  //         duration: 3000,
+  //         verticalPosition: 'bottom',
+  //         horizontalPosition: 'center'
+  //       });
+  //     }
+  //   })
+  // }
+
   deleteCategory(id: any) {
-    this.blogsService.deleteCategoryList(id).subscribe({
-      next: (res: any) => {
-        this._snackBar.open(res.message, 'Close', {
-          duration: 3000,
-          verticalPosition: 'bottom',
-          horizontalPosition: 'center'
-        });
-        if (res.status === 200) {
-          this.getCategoryList()
-        }
-      }, error: (err: HttpErrorResponse) => {
-        this._snackBar.open(err.statusText, 'Close', {
-          duration: 3000,
-          verticalPosition: 'bottom',
-          horizontalPosition: 'center'
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to undo this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.blogsService.deleteCategoryList(id).subscribe({
+          next: (res: any) => {
+            this._snackBar.open(res.message, 'Close', {
+              duration: 3000,
+              verticalPosition: 'bottom',
+              horizontalPosition: 'center'
+            });
+            if (res.status === 200) {
+              this.getCategoryList();
+            }
+          },
+          error: (err: HttpErrorResponse) => {
+            this._snackBar.open(err.statusText, 'Close', {
+              duration: 3000,
+              verticalPosition: 'bottom',
+              horizontalPosition: 'center'
+            });
+          }
         });
       }
-    })
+    });
   }
 }
